move components to src
