+++ src/components/ScriptAdder/index.js
